requirements.txt
pytest
playwright
python-dotenv
pytest-html
pytest-xdist
allure-pytest
